backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

pydantic==1.10.24

sqlalchemy==2.0.40
alembic==1.12.1

python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pdfplumber==0.10.3
python-docx==1.1.0
python-magic==0.4.27
huggingface-hub==0.24.6
transformers==4.35.2
torch==2.9.1
sentence-transformers==2.2.2

numpy>=2.0.0
pydub==0.25.1
